Ported more of the template from LateX to Typst.
The thm-helpers maps faculties and other things to objects containing
their descriptions in german and english just like in the LateX
template. Additionally some work towards porting chapter 1 and 5 has
been done. One big issue is that the total number of pages doesn't work
with counter, because we reset the counter when swapping between roman
page numbers and back. There is also a lot of layout finetuning to be
done.